# docs/swagger.yaml
basePath: /
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        type: boolean
      message:
        type: string
    type: object
  handler.ListOrderResponse:
    properties:
      message:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      page:
        type: integer
      size:
        type: integer
      total_orders:
        type: integer
      total_pages:
        type: integer
    type: object
...
  handler.OrderRequest:
    properties:
      order_reference:
        type: string
      products:
        items:
          $ref: '#/definitions/handler.ProductDTO'
        type: array
      user_id:
        type: string
    required:
    - order_reference
    - products
    - user_id
    type: object
  handler.OrderResponse:
    properties:
      message:
        type: string
      order:
        $ref: '#/definitions/model.Order'
    type: object
  handler.ProductDTO:
    properties:
      product_code:
        type: string
      product_quantity:
        type: integer
    required:
    - product_code
    - product_quantity
    type: object
...
  handler.UpdateOrderRequest:
    properties:
      order_status:
        type: string
    required:
    - order_status
    type: object
...
  model.Order:
    properties:
      created_at:
        type: string
      order_reference:
        example: order123
        type: string
      order_status:
        allOf:
        - $ref: '#/definitions/model.OrderStatus'
        description: Pending, Shipped, Delivered, Canceled
        example: Pending
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      total_price:
        example: 10.5
        type: number
      updated_at:
        type: string
    type: object
  model.OrderStatus:
    enum:
    - Pending
    - Shipped
    - Delivered
    - Cancelled
    type: string
    x-enum-varnames:
    - Pending
    - Shipped
    - Delivered
    - Cancelled
  model.Product:
    properties:
      created_at:
        type: string
      currency:
        type: string
      price:
        type: number
      product_code:
        type: string
      product_description:
        type: string
      product_name:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is a sample server Instashop server.
  title: Instashop Swagger API
  version: "1.0"
paths:
  /api/v1/admin/order/{order_reference}/status:
    put:
      description: Updates the status of a specific order for a user
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order Reference
        in: path
        name: order_reference
        required: true
        type: string
      - description: Update Status Request
        in: body
        name: updateOrder
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order status updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/handler.OrderResponse'
            - properties:
                ' order':
                  $ref: '#/definitions/model.Order'
                message:
                  type: string
              type: object
        "400":
          description: Order has been (Shipped | Delivered)
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - Orders
...
  /api/v1/user/order:
    get:
      description: Retrieves all orders for a given user with an optional status filter
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Order Status (Pending, Shipped, Delivered, Cancelled)
        in: query
        name: order_status
        type: string
      - description: Page (Default 1)
        in: query
        name: page
        type: string
      - description: Size (Default 10)
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user orders
          schema:
            allOf:
            - $ref: '#/definitions/handler.ListOrderResponse'
            - properties:
                ' message':
                  type: string
                ' page':
                  type: integer
                ' size':
                  type: integer
                ' total_orders':
                  type: integer
                ' total_pages':
                  type: integer
                orders:
                  items:
                    $ref: '#/definitions/model.Order'
                  type: array
              type: object
        "500":
          description: Failed to retrieve orders
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Get user orders
      tags:
      - Orders
    post:
      description: Creates a new order for a user with a list of products
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order Request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.OrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order placed successfully
          schema:
            allOf:
            - $ref: '#/definitions/handler.OrderResponse'
            - properties:
                ' order':
                  $ref: '#/definitions/model.Order'
                message:
                  type: string
              type: object
        "400":
          description: Invalid input
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "404":
          description: User not found
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "500":
          description: Failed to place order
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Place a new order
      tags:
      - Orders
  /api/v1/user/order/{order_reference}/cancel:
    put:
      description: Cancels an order associated with the provided order reference
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Authenticated User ID
        in: header
        name: user_id
        required: true
        type: string
      - description: Order Reference
        in: path
        name: order_reference
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order cancelled successfully
          schema:
            allOf:
            - $ref: '#/definitions/handler.OrderResponse'
            - properties:
                ' order':
                  $ref: '#/definitions/model.Order'
                message:
                  type: string
              type: object
        "400":
          description: Invalid input or Order in %s status cannot be cancelled
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "401":
          description: Unauthorized access
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "404":
          description: Order not found
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "500":
          description: Failed to update order
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Cancel a user order
      tags:
      - Orders
...
swagger: "2.0"